# app/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Souk API",
        "description": "API du marketplace",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur local"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/tags/create": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Créer un tag",
                "description": "Permet de créer un nouveau tag avec nom, couleur et description.",
                "operationId": "b531dede33b029680edfb5b438c4ef3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Name"
                                ],
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Promotion"
                                    },
                                    "Color": {
                                        "type": "string",
                                        "example": "#FF5733",
                                        "nullable": true
                                    },
                                    "Description": {
                                        "type": "string",
                                        "example": "Produits en promotion",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag créé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Un tag avec ce nom existe déjà"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/create-by-name": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Créer un tag à partir du nom uniquement",
                "description": "Crée un tag en ne fournissant que le nom ; Color, Description et IsActive utilisent leurs valeurs par défaut. Retourne l'ID du tag créé.",
                "operationId": "9cda73135e6a13862044f163cc5c0857",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Name"
                                ],
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "Nouveauté"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tag créé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 42
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Un tag avec ce nom existe déjà"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Liste des tags (avec filtres, recherche, pagination)",
                "operationId": "4d0f4baf9e6fa992c0d85fbf140ebf17",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Name",
                                "CreatedAt",
                                "IsActive"
                            ]
                        }
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
        "/api/tags/{id}/disable": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Désactiver un tag",
                "description": "Met à jour le statut d'un tag existant en le désactivant (IsActive = 0).",
                "operationId": "b2a7c39ebf4c0f9585a146bfd4618790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 42
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag désactivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tag désactivé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "isActive": {
                                                    "type": "boolean",
                                                    "example": false
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Tag introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrer uniquement votre JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}